* ext/tk/tkutil/tkutil.c (cbsubst_info): fix shorten-64-to-32 warnings.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@33800 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/tk/tkutil/tkutil.c
@@ -1135,7 +1135,7 @@ tcl2rb_num_or_nil(self, value)
 
 #define CBSUBST_TBL_MAX (256)
 struct cbsubst_info {
-    int   full_subst_length;
+    long  full_subst_length;
     long  keylen[CBSUBST_TBL_MAX];
     char  *key[CBSUBST_TBL_MAX];
     char  type[CBSUBST_TBL_MAX];
